refactor(OuterWrapper): Make reusable component
