+++ packages/react-component-library/src/styled-components/index.ts
@@ -1,2 +1,3 @@
-export * from './input'
+export * from './partials/StyledLabel'
+export * from './partials/StyledOuterWrapper'
 export * from './GlobalStyle'
